Rename base64sm to b64
